requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pytesseract>=0.3.10
pure-python-adb>=0.3.0
yt-dlp>=2023.10.0
pyyaml>=6.0
numpy>=1.24.0
pillow>=10.0.0
tensorboard>=2.14.0
pytest>=7.4.0
tqdm>=4.66.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
